fix: fix: Room의 price 중 최소 가격을 Accommodation의 price 컬럼에 전달 - rate를 전달하는 방식과 똑같이 구현 - repository 단위테스트 완료 - swagger-ui 테스트 완료

+++ src/main/java/com/core/miniproject/src/room/controller/RoomPublicController.java
@@ -21,9 +21,9 @@ public class RoomPublicController {
     private final RoomService roomService;
 
     @GetMapping("/v1/accommodation/{accommodationId}")
-    public BaseResponse<List<RoomResponse>> findAllRoomByAccommodationId(@PathVariable("accommodationId") int accommodationId){
+    public BaseResponse<List<RoomResponse>> findAllRoomByAccommodationId(@PathVariable("accommodationId") Long accommodationId){
 
-        List<RoomResponse> roomResponses = roomService.findAllRoomByAccommodationId((long)accommodationId);
+        List<RoomResponse> roomResponses = roomService.findAllRoomByAccommodationId(accommodationId);
 
         return BaseResponse.response(roomResponses);
     }
